+++ test/controllers/async/parent_controller_test.rb
@@ -22,12 +22,12 @@ class Async::ParentControllerTest < ActionController::TestCase
 
   test 'should start new reply on forum topic for public user' do
     post :reply, params: {
-      chapter_id: chapters(:one).to_param,
+      topic_id: topics(:one).to_param,
       parent_reply_id: 'root', reply_id: 'new'
     }, format: 'js'
     assert_not_nil assigns(:reply)
     assert_equal true, assigns(:reply).new_record?
-    assert_equal Chapter, assigns(:reply).parent.class
+    assert_equal Topic, assigns(:reply).parent.class
     assert_template 'reply'
     assert_response :success
   end
@@ -45,7 +45,7 @@ class Async::ParentControllerTest < ActionController::TestCase
   test 'should start new reply on forum topic for regular user' do
     login(@regular_user)
     post :reply, params: {
-      chapter_id: chapters(:one).to_param, parent_reply_id: 'root', reply_id: 'new'
+      topic_id: topics(:one).to_param, parent_reply_id: 'root', reply_id: 'new'
     }, format: 'js'
     assert_template 'reply'
     assert_response :success
@@ -66,11 +66,11 @@ class Async::ParentControllerTest < ActionController::TestCase
   test 'should sign in as user when replying to forum topic' do
     post :login, params: {
       email: 'user_1@example.com', password: 'password',
-      chapter_id: chapters(:one).to_param
+      topic_id: topics(:one).to_param
     }, format: 'js'
     assert_not_nil assigns(:reply)
     assert_equal true, assigns(:reply).new_record?
-    assert_equal Chapter, assigns(:reply).parent.class
+    assert_equal Topic, assigns(:reply).parent.class
     assert_template 'create'
     assert_response :success
   end
@@ -87,7 +87,7 @@ class Async::ParentControllerTest < ActionController::TestCase
   test 'should not sign in as user with incorrect password when reply to forum topic' do
     post :login, params: {
       email: 'user_1@example.com', password: 'wrong',
-      chapter_id: chapters(:one).to_param
+      topic_id: topics(:one).to_param
     }, format: 'js'
     assert_template 'new'
     assert_response :success